Move parsers to a different file
